*** rama-maximo : Se estructura el proyecto
> .
> Se creo una carpeta para componentes y accione
> en el controlador y vistas -> panel
