+++ Veontec/src/vista/paneles/acciones/PanelCrearProducto.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package vista.paneles;
+package vista.paneles.acciones;
 
 import src.Recursos;
