+++ Veontec/src/vista/paneles/acciones/PanelHacerCompra.form
