+++ Veontec/src/vista/paneles/componentes/PanelCardProyectos.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package vista.paneles;
+package vista.paneles.componentes;
 
 import index.Veontec;
 import java.awt.Color;
